Move Hubverse nowcast into data pipeline
