+++ pipelines/epiautogp/__init__.py
@@ -2,8 +2,8 @@
 EpiAutoGP integration module for cfa-stf-routine-forecasting pipelines.
 """
 
+from pipelines.data.hubverse_nowcast import HubverseNowcast
 from pipelines.epiautogp.epiautogp_forecast_utils import setup_forecast_pipeline
-from pipelines.epiautogp.hubverse_nowcast import HubverseNowcast
 from pipelines.epiautogp.nowcast import FixedNowcast, NowcastData, NowcastSource
 from pipelines.epiautogp.prep_epiautogp_data import convert_to_epiautogp_json
 from pipelines.epiautogp.reporting_delay import (
