+++ pipelines/tests/data/test_hubverse_nowcast.py
@@ -5,9 +5,9 @@ import datetime as dt
 import polars as pl
 import pytest
 
+from pipelines.data.hubverse_nowcast import HubverseNowcast
 from pipelines.epiautogp.epiautogp_forecast_utils import _resolve_nowcast_source
 from pipelines.epiautogp.forecast_spec import ForecastSpec
-from pipelines.epiautogp.hubverse_nowcast import HubverseNowcast
 from pipelines.epiautogp.nowcast import NowcastData
 
 ORIGIN = dt.date(2026, 7, 18)
