+++ pipelines/tests/epiautogp/test_prep_epiautogp_data.py
@@ -14,12 +14,12 @@ from unittest.mock import MagicMock
 import polars as pl
 import pytest
 
+from pipelines.data.hubverse_nowcast import HubverseNowcast
 from pipelines.epiautogp.epiautogp_forecast_utils import (
     ForecastPipelineContext,
     ForecastSpec,
     ModelPaths,
 )
-from pipelines.epiautogp.hubverse_nowcast import HubverseNowcast
 from pipelines.epiautogp.nowcast import NowcastData
 from pipelines.epiautogp.prep_epiautogp_data import (
     _apply_date_exclusions,
